added queries for webui.
