+++ src/queries/getLoginUser.ts
@@ -1,0 +1,13 @@
+import { prisma } from '../prismaClient'
+import { Profile } from '../../src/models/profile'
+
+export const getLoginUser = async (req: any) => {
+  const profile = <Profile>req.user
+  if (profile === undefined || profile === null) {
+    throw new Error('login user not found in request')
+  }
+  const loginUser = await prisma.user.findUnique({
+    where: { id: profile.id },
+  })
+  return loginUser
+}
