+++ views/dashboard.ejs
@@ -37,7 +37,7 @@
                             <% } %>
                         </div>
                     </div>
-                    <% rulesRatingMap.forEach((value, key) => { %>
+                    <% powerGraphData.forEach((value, key) => { %>
                         <% if (0 < value.data.labels.length) { %>
                             <div class="row">
                                 <div class="col-12">
