+++ views/ranking.ejs
@@ -15,15 +15,15 @@
                             <p class="text-primary m-0 fw-bold">Ranking</p>
                         </div>
                         <div class="card-body">
-                            <% if (rankingMap) { %>
+                            <% if (rankingData) { %>
                                 <div class="row my-1">
                                     <div class="col-3">ルール</div>
                                     <div class="col-3">パワーランキング</div>
                                     <div class="col-3">計測パワーランキング</div>
                                 </div>
                                 <% rules.forEach((rule) => {%>
-                                    <% if (rankingMap.has(rule.code)) { %>
-                                        <% const ruleRanking = rankingMap.get(rule.code) %>
+                                    <% if (rankingData.has(rule.code)) { %>
+                                        <% const ruleRanking = rankingData.get(rule.code) %>
                                         <div class="row my-1">
                                             <div class="col-3 d-flex align-items-center">
                                                 <h4 class="small fw-bold"><%= rule.name %></h4>
